Added support for -gridView:willDisplayCell:forItemAtIndexPath: @optional delegate method (#20)
In the course of this, I also refactored the code for loading a cell from the
datasource into the _visibleCells array and setting its frame into
-_loadCellAtVisibleIndexPath:.

Finally, I refactored the display of cell views into -_displayCell:atIndexPath:.
This informs the delegate that the cell is about to be displayed, and then adds
it as a subview and sends it to the back. By the time this is called, the cell
should have the proper frame, etc.