Fixed KKGridView to work with Xcode workspaces properly
The problem: it's a real pain to get it to work, and it's not clear how to get
Xcode to recognize the headers. And then once you figure that out, you realize
you're fucked the moment you want to do a build and archive. So, I followed
these brilliant instructions:
http://mrox.net/blog/2011/06/19/framework-like-static-libraries/.

Now, shit just works. You don't need any special header paths or anything in
your main project. That's all handled in the build settings for KKGridView.
Archive builds work. From now on, in external projects, just import KKGridView
as though it were a framework:

    #import <KKGridView/KKGridView.h>

Additionally, I made Definitions.h a public header since it definitely needs to
be exposed.