Added support for temporary highlights like UITableView
The expected behavior is that when you touch down on a cell, it is
highlighted, and that touch can be cancelled in the normal way; if you
release your touch without moving your finger off of the cell, then
you've selected it. This is how things work in UITableView, and this is
how it works in KKGridView now too.

It's all done with a single gesture recognizer; this seems to work, but
it's a little complex. There's a little bit of wacky stuff that had to
be done to make this interact in the right way with the panning gesture
recognizer of the scrollview. I'm definitely open to superior ways of
dealing with this.