Close #81
So, let's say we the consumer opts to not provide section headers or
footers (neither views nor titles). Obviously, we want some space
between sections by default (unless they implement the
`-gridView:heightFor{Header|Footer}InSection:` methods). If we put the
default space the header, we get an empty spot before the first section,
which is bad. If we put the space in the footer, it all works out.

So, the behavior here is that footers (whether or not they are empty)
are give 25.0px of space, and headers are given none. If a title or a
view is provided, then headers are also given 25.0px. If other behavior
is desired, consumers may implement the appropriate dataSource methods.

It may be logical at some point to make a call to
`-gridView:viewFor{Header|Footer}InSection:` to get the actual height of
the provided view into the metrics cache; this way, the consumer doesn't
need to implement the height callbacks. I'll leave that for a future
commit.