Removed -_heightForSection: outright
There was only one place where it was reasonable to call that method,
namely in -reloadContentSize when the _metrics are being updated. In
all other cases, section height should be gotten from _metrics.

To prevent us from accidentally using this method, I've removed it and
reintegrated its code into -reloadContentSize.