Cleanup switch statements in -_layoutAccessories
Switch statements are great if we want to perform an effect depending
on the value of something. However, in these instances, the goal was
basically a function (KKGridViewAccessoryPosition → CGPoint). This can
be encoded very nicely using an array of CGPoints indexed by
KKGridViewAccessoryPosition values: just think of it as a function
that is invoked with the subscript ([]) operator.

Using this technique in combination with C99's fancy initializers, we
get more compact, declarative code.